DNS: package refactoring
