+++ DnsService/src/main/scala/org/gammf/proxima/dns/hierarchy/actors/DNSNodeActor.scala
@@ -1,10 +1,10 @@
-package org.gammf.proxima.dns.actors
+package org.gammf.proxima.dns.hierarchy.actors
 
 import akka.actor.{ActorRef, Props}
-import org.gammf.proxima.dns.messages._
-import org.gammf.proxima.dns.util.Role._
-import org.gammf.proxima.dns.util.Service.StringService
-import org.gammf.proxima.dns.util.ServiceAddress
+import org.gammf.proxima.dns.hierarchy.messages._
+import org.gammf.proxima.dns.hierarchy.util.Role._
+import org.gammf.proxima.dns.hierarchy.util.Service.StringService
+import org.gammf.proxima.dns.hierarchy.util.ServiceAddress
 
 /**
   * Represents a simple DNS node identified by a name and a role, offering a [[StringService]].
